Retry deleting chrome apps dir in OsIntegrationTestOverrideImpl.
Deletion during test teardown sometimes flakily fails because the
task that sets the icon for the chrome apps dir happens to be
running while the directory is being deleted, causing deletion to
fail. Adding a single retry is the simplest way to work around
this flakiness. If this turns out to not work for some reason we
can consider more complicated schemes to wait for or skip setting
the icon.

Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5626732
Commit-Queue: Marijn Kruisselbrink <mek@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1314212}

+++ chrome/browser/web_applications/test/os_integration_test_override_impl.cc
@@ -281,7 +281,15 @@ bool OsIntegrationTestOverrideImpl::SimulateDeleteShortcutsByUser(
 #if BUILDFLAG(IS_MAC)
 bool OsIntegrationTestOverrideImpl::DeleteChromeAppsDir() {
   if (chrome_apps_folder_.IsValid()) {
-    return chrome_apps_folder_.Delete();
+    bool success = chrome_apps_folder_.Delete();
+    if (!success) {
+      // Creating shortcuts kicks of an asynchronous task to eventually update
+      // the icon of `chrome_apps_folder_`. If that task happens to run during
+      // the above Delete() call deletion might fail. If that is the case, a
+      // single retry should be enough to be able to delete the folder anyway.
+      success = chrome_apps_folder_.Delete();
+    }
+    return success;
   } else {
     return false;
   }
@@ -818,7 +826,7 @@ OsIntegrationTestOverrideImpl::~OsIntegrationTestOverrideImpl() {
   EXPECT_TRUE(!quick_launch_.IsValid() || quick_launch_.Delete());
   EXPECT_TRUE(!startup_.IsValid() || startup_.Delete());
 #elif BUILDFLAG(IS_MAC)
-  EXPECT_TRUE(!chrome_apps_folder_.IsValid() || chrome_apps_folder_.Delete());
+  EXPECT_TRUE(!chrome_apps_folder_.IsValid() || DeleteChromeAppsDir());
 #elif BUILDFLAG(IS_LINUX)
   EXPECT_TRUE(!desktop_.IsValid() || desktop_.Delete());
   EXPECT_TRUE(!startup_.IsValid() || startup_.Delete());
